Massive code refactor, added new global properties that can be set in the application.yaml file.
